ci: update ruff pre-commit
Change id, as id used was a legacy alias relating to an older version of ruff-pre-commit. Also change args for this reason.
Additionally, move ruff-pre-commit to before black and isort, following package recommendations. Specify to only use ruff as a linter, not also as a formatter as black is being used as a formatter.